Move user content out of dashboard into library
