+++ applications/dashboard/src/scripts/components/PageLoading.tsx
@@ -1,3 +1,8 @@
+/**
+ * @copyright 2009-2018 Vanilla Forums Inc.
+ * @license https://opensource.org/licenses/GPL-2.0 GPL-2.0
+ */
+
 import * as React from "react";
 import { t } from "@library/application";
 import { ILoadable, LoadStatus } from "@library/@types/api";
