--- applications/dashboard/src/scripts/entries/forum.ts
@@ -1,10 +1,0 @@
-/**
- * @copyright 2009-2018 Vanilla Forums Inc.
- * @license https://opensource.org/licenses/GPL-2.0 GPL-2.0
- */
-
-import "@dashboard/app/index";
-
-// The forum section needs these legacy scripts that have been moved into the bundled JS so it could be refactored.
-// Other sections should not need this yet.
-import "@dashboard/legacy";
