+++ applications/dashboard/src/scripts/pages/authenticate/authenticatorsActions.ts
