+++ applications/dashboard/src/scripts/pages/authenticate/authenticatorsReducer.ts
@@ -3,38 +3,36 @@
  * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
  */
 
-import * as actions from "@dashboard/state/users/requestPasswordActions";
-import { IRequestPasswordState } from "@dashboard/@types/state";
+import * as actions from "@dashboard/pages/authenticate/authenticatorsActions";
+import { IAuthenticatorState } from "@dashboard/@types/state";
 import { LoadStatus } from "@library/@types/api";
 
-const initialState: IRequestPasswordState = {
+const initialState: IAuthenticatorState = {
     status: LoadStatus.PENDING,
 };
 
-export default function passwordReducer(
-    state: IRequestPasswordState = initialState,
+export default function authenticatorsReducer(
+    state: IAuthenticatorState = initialState,
     action: actions.ActionTypes,
-): IRequestPasswordState {
+): IAuthenticatorState {
     switch (action.type) {
-        case actions.POST_REQUEST_PASSWORD_REQUEST:
+        case actions.GET_USER_AUTHENTICATORS_REQUEST:
             return {
                 ...state,
                 status: LoadStatus.LOADING,
             };
-        case actions.POST_REQUEST_PASSWORD_SUCCESS:
+        case actions.GET_USER_AUTHENTICATORS_SUCCESS:
             return {
                 status: LoadStatus.SUCCESS,
                 data: action.payload.data,
                 error: undefined,
             };
-        case actions.POST_REQUEST_PASSWORD_ERROR:
+        case actions.GET_USER_AUTHENTICATORS_ERROR:
             return {
                 ...state,
                 status: LoadStatus.ERROR,
                 error: action.payload,
             };
-        case actions.AFTER_REQUEST_PASSWORD_SUCCESS_NAVIGATE:
-            return initialState;
         default:
             return state;
     }
