+++ applications/dashboard/src/scripts/pages/authenticate/components/RememberPasswordLink.tsx
