+++ applications/dashboard/src/scripts/pages/authenticate/components/SSOMethods.tsx
