+++ applications/dashboard/src/scripts/pages/authenticate/passwordActions.ts
