+++ applications/dashboard/src/scripts/pages/authenticate/passwordReducer.ts
@@ -3,31 +3,31 @@
  * @license http://www.opensource.org/licenses/gpl-2.0.php GNU GPL v2
  */
 
-import * as actions from "@dashboard/state/authenticate/authenticatorsActions";
-import { IAuthenticatorState } from "@dashboard/@types/state";
+import * as actions from "@dashboard/pages/authenticate/passwordActions";
+import { IPasswordState } from "@dashboard/@types/state";
 import { LoadStatus } from "@library/@types/api";
 
-const initialState: IAuthenticatorState = {
+const initialState: IPasswordState = {
     status: LoadStatus.PENDING,
 };
 
-export default function authenticatorsReducer(
-    state: IAuthenticatorState = initialState,
+export default function passwordReducer(
+    state: IPasswordState = initialState,
     action: actions.ActionTypes,
-): IAuthenticatorState {
+): IPasswordState {
     switch (action.type) {
-        case actions.GET_USER_AUTHENTICATORS_REQUEST:
+        case actions.POST_AUTHENTICATE_PASSWORD_REQUEST:
             return {
                 ...state,
                 status: LoadStatus.LOADING,
             };
-        case actions.GET_USER_AUTHENTICATORS_SUCCESS:
+        case actions.POST_AUTHENTICATE_PASSWORD_SUCCESS:
             return {
                 status: LoadStatus.SUCCESS,
                 data: action.payload.data,
                 error: undefined,
             };
-        case actions.GET_USER_AUTHENTICATORS_ERROR:
+        case actions.POST_AUTHENTICATE_PASSWORD_ERROR:
             return {
                 ...state,
                 status: LoadStatus.ERROR,
