+++ library/src/scripts/user-content/CollapsableContent.tsx
