+++ library/src/scripts/user-content/embeds/getty.ts
