+++ library/src/scripts/user-content/embeds/quote.tsx
@@ -7,7 +7,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 import { registerEmbedComponent, IEmbedProps, IEmbedData, IQuoteEmbedData } from "@library/embeds";
 import { onContent, t, formatUrl, makeProfileUrl } from "@library/application";
-import CollapsableUserContent from "@dashboard/app/user-content/CollapsableContent";
+import CollapsableUserContent from "@library/user-content/CollapsableContent";
 import uniqueId from "lodash/uniqueId";
 import classnames from "classnames";
 import api from "@library/apiv2";
