+++ library/src/scripts/user-content/emoji.ts
