+++ plugins/rich-editor/src/scripts/state/mention/MentionTrie.test.ts
@@ -6,7 +6,7 @@
 
 import MentionTrie, { IMentionNode, IMentionValue } from "./MentionTrie";
 import { expect } from "chai";
-import { LoadStatus } from "@dashboard/@types/api";
+import { LoadStatus } from "@library/@types/api";
 
 const LOADING_VALUE: IMentionValue = {
     status: LoadStatus.LOADING,
